Fixes the RConsole.
