+++ r-console/pom.xml
@@ -1,15 +1,16 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <artifactId>r-console</artifactId>
+    <name>R console</name>
+    <version>5.1.0</version>
+    <url>http://orbisgis.org</url>
     <parent>
         <groupId>org.orbisgis</groupId>
         <artifactId>orbisgis-view-plugins</artifactId>
         <version>5.1</version>
+        <relativePath>../</relativePath>
     </parent>
-    <name>R console</name>
-    <version>5.1.0</version>
-    <url>http://orbisgis.org</url>
-    <artifactId>r-console</artifactId>
     <description>A R console for OrbisGIS platform based on Renjin.</description>
     <packaging>bundle</packaging>
     <build>
@@ -38,7 +39,7 @@
                         <!--                        <Import-Package>groovy.sql;version=2.1.9,*</Import-Package>-->
                         <Bundle-Vendor>LabSTICC - CNRS, UMR 6285.</Bundle-Vendor>
                         <Bundle-Category>OrbisGIS, Console, Language</Bundle-Category>
-                        <Bundle-Icon>org/orbisgis/r/console/icon.png;size=32</Bundle-Icon>
+                        <Bundle-Icon>org/orbisgis/r/icon.png;size=32</Bundle-Icon>
                     </instructions>
                 </configuration>
             </plugin>
@@ -56,30 +57,12 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <poDirectory>${project.build.sourceDirectory}/../resources/org/orbisgis/r/console/translation/</poDirectory>
-                    <targetBundle>org.orbisgis.r.console.Messages</targetBundle>
+                    <poDirectory>${project.build.sourceDirectory}/../resources/org/orbisgis/r/translation/</poDirectory>
+                    <targetBundle>org.orbisgis.r.Messages</targetBundle>
                     <keywords>-ktr</keywords>
                     <outputFormat>properties</outputFormat>
                 </configuration>
             </plugin>
-<!--<plugin>
-            <artifactId>maven-assembly-plugin</artifactId>
-            <version>2.3</version>
-            <configuration>
-                <descriptorRefs>
-                    <descriptorRef>jar-with-dependencies</descriptorRef>
-                </descriptorRefs>
-            </configuration>
-            <executions>
-                <execution>
-                    <id>make-assembly</id>  this is used for inheritance merges 
-                    <phase>package</phase>   bind to the packaging phase 
-                    <goals>
-                        <goal>single</goal>
-                    </goals>
-                </execution>
-            </executions>
-        </plugin>-->
         </plugins>
     </build>
     <properties>
@@ -105,18 +88,13 @@
         <dependency>
             <groupId>org.renjin</groupId>
             <artifactId>renjin-script-engine</artifactId>
-            <version>0.8.1615</version>
-        </dependency>        
+            <version>0.9.0-SNAPSHOT</version>
+        </dependency>
         <dependency>
             <groupId>org.osgi</groupId>
             <artifactId>org.osgi.enterprise</artifactId>
             <version>5.0.0</version>
             <scope>compile</scope>
-        </dependency>        
-        <dependency>
-            <groupId>org.orbisgis</groupId>
-            <artifactId>map-editor-api</artifactId>
-            <version>5.1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
     <repositories>
