+++ r-console/src/main/resources/org/orbisgis/r/i18n.properties
