+++ r-console/src/main/resources/org/orbisgis/r/icon.png
