+++ r-console/src/main/resources/org/orbisgis/r/icons/erase.png
