+++ r-console/src/main/resources/org/orbisgis/r/icons/open.png
